change life from int to float

+++ core/src/no/fusiontd/components/Durability.java
@@ -3,9 +3,9 @@ package no.fusiontd.components;
 import com.badlogic.ashley.core.Component;
 
 public class Durability implements Component {
-    public int life = 1;
+    public float life = 1;
 
-    public Durability(int maxLife) {
+    public Durability(float maxLife) {
         this.life = maxLife;
     }
